Clippy

+++ src/frontend/chat_mode/on_key.rs
@@ -45,13 +45,13 @@ pub fn on_key(app: &mut App, key: KeyEvent) {
             app.vertical_scroll = app.vertical_scroll.saturating_add(1);
             app.vertical_scroll_state = app
                 .vertical_scroll_state
-                .position(app.vertical_scroll as usize);
+                .position(app.vertical_scroll);
         }
         KeyCode::PageUp => {
             app.vertical_scroll = app.vertical_scroll.saturating_sub(1);
             app.vertical_scroll_state = app
                 .vertical_scroll_state
-                .position(app.vertical_scroll as usize);
+                .position(app.vertical_scroll);
         }
         _ => {
             app.text_input.input(key);
